requirements.txt
# Project Cerberus Dependencies
# Simple, minimal dependencies for a 3rd-year student project

fastapi==0.104.1          # Modern web framework for building APIs
uvicorn==0.24.0           # ASGI server to run FastAPI
python-dotenv==1.0.0      # Load environment variables from .env file
google-generativeai==0.3.1  # Gemini API for AI interactions
pydantic==2.5.0           # Data validation (comes with FastAPI but explicit is better)
